Major Improvment
Added the custom cursor and started to bring everything together. The editor support HTML and is using it by default. This is basic HTML and not everything is being highlighted. You cannot select and using the mouse, arrow keys must be used at the moment. working on getting that all working and adding more language support. Then I'll be turning it into a library for ease of use.